finally combined community with case studies
