--- bifrost/app/projects/layout.tsx
@@ -1,9 +1,0 @@
-import { Layout } from "@/app/components/Layout";
-
-export default function RootLayount({
-    children,
-}: {
-    children: React.ReactNode;
-}) {
-    return <Layout>{children}</Layout>;
-}
